pyproject.toml
[tool.poetry]
name = "ai-news-letter"
version = "0.1.0"
description = "AI News Letter" 
authors = ["Hein"]

[tool.poetry.dependencies]
python = "^3.10.11"
crewai = "^0.11.0"
duckduckgo-search = "^4.4.2"
pyowm = '3.3.0'
python-dotenv = "^1.0.1"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.3"
langchain-community = "0.0.10"
langchain-core = "0.1.8"
openai = "1.23.1"

[tool.pyright]
# https://github.com/micorsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E','W','F','I','B','C4','ARG','SIM']
ignore = ['W291','W292','W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
